starting to breakup libusb
