+++ src/thunderbots/software/radio_communication/mrf/libusb.h
@@ -14,6 +14,9 @@
 #include <unordered_map>
 #include <vector>
 
+#include "libusb_errors.h"
+#include "libusb_device.h"
+
 #include "async_operation.h"
 #include "util/noncopyable.h"
 
@@ -28,82 +31,6 @@ namespace USB
         void usb_context_pollfd_remove_trampoline(int fd, void *user_data);
         void usb_transfer_handle_completed_transfer_trampoline(libusb_transfer *transfer);
     }
-
-    /**
-     * An error that occurs in a libusb library function.
-     */
-    class Error : public std::runtime_error
-    {
-       public:
-        /**
-         * Constructs a new error object with a message.
-         *
-         * @param[in] msg a detail message explaining the error
-         */
-        explicit Error(const std::string &msg);
-    };
-
-    /**
-     * An error that occurs on a USB transfer.
-     */
-    class TransferError : public Error
-    {
-       public:
-        /**
-         * Constructs a new transfer error.
-         *
-         * @param[in] endpoint the endpoint number, with bit 7 used to indicate
-         * direction
-         *
-         * @param[in] msg a detail message explaining the error (UTF-8 encoded)
-         */
-        explicit TransferError(unsigned int endpoint, const char *msg);
-    };
-
-    /**
-     * An error that occurs when a USB transfer times out.
-     */
-    class TransferTimeoutError final : public TransferError
-    {
-       public:
-        /**
-         * Constructs a new timeout error object.
-         *
-         * @param[in] endpoint the endpoint number, with bit 7 used to indicate
-         * direction
-         */
-        explicit TransferTimeoutError(unsigned int endpoint);
-    };
-
-    /**
-     * An error that occurs when a USB stall occurs.
-     */
-    class TransferStallError final : public TransferError
-    {
-       public:
-        /**
-         * Constructs a new stall error object.
-         *
-         * @param[in] endpoint the endpoint number, with bit 7 used to indicate
-         * direction
-         */
-        explicit TransferStallError(unsigned int endpoint);
-    };
-
-    /**
-     * An error that occurs when a USB transfer is cancelled.
-     */
-    class TransferCancelledError final : public TransferError
-    {
-       public:
-        /**
-         * Constructs a new cancelled transfer error object.
-         *
-         * @param[in] endpoint the endpoint number, with bit 7 used to indicate
-         * direction
-         */
-        explicit TransferCancelledError(unsigned int endpoint);
-    };
 
     /**
      * A libusb context.
@@ -131,115 +58,6 @@ namespace USB
         friend class DeviceHandle;
 
         libusb_context *context;
-    };
-
-    /**
-     * A collection of information about a USB device.
-     */
-    class Device final
-    {
-       public:
-        /**
-         * Makes a copy of a device record.
-         *
-         * @param[in] copyref the object to copy
-         */
-        Device(const Device &copyref);
-
-        /**
-         * Destroys the device information record.
-         */
-        ~Device();
-
-        /**
-         * Assigns a device information record.
-         *
-         * @param[in] assgref the object to copy from
-         *
-         * @return this object
-         */
-        Device &operator=(const Device &assgref);
-
-        /**
-         * Returns the 16-bit vendor ID from the device’s device descriptor.
-         *
-         * @return the vendor ID
-         */
-        unsigned int vendor_id() const
-        {
-            return device_descriptor.idVendor;
-        }
-
-        /**
-         * Returns the 16-bit product ID from the device’s device descriptor.
-         *
-         * @return the product ID
-         */
-        unsigned int product_id() const
-        {
-            return device_descriptor.idProduct;
-        }
-
-        /**
-         * Returns the serial number from the device’s device and string
-         * descriptors.
-         *
-         * @return the serial number, or an empty string if the device does not
-         * expose a serial number
-         */
-        std::string serial_number() const;
-
-       private:
-        friend class DeviceList;
-        friend class DeviceHandle;
-
-        libusb_context *context;
-        libusb_device *device;
-        libusb_device_descriptor device_descriptor;
-
-        explicit Device(libusb_device *device);
-    };
-
-    /**
-     * A list of USB devices.
-     */
-    class DeviceList final : public NonCopyable
-    {
-       public:
-        /**
-         * Constructs a list of all USB devices attached to the system.
-         *
-         * @param[in] context the library context in which to operate
-         */
-        explicit DeviceList(Context &context);
-
-        /**
-         * Frees the list of devices.
-         */
-        ~DeviceList();
-
-        /**
-         * Returns the size of the list.
-         *
-         * @return the number of devices in the list
-         */
-        std::size_t size() const
-        {
-            return size_;
-        }
-
-        /**
-         * Returns a device from the list.
-         *
-         * @param[in] i the index of the device to return, counting from zero
-         *
-         * @return the device
-         */
-        Device operator[](const std::size_t i) const;
-
-       private:
-        std::size_t size_;
-        libusb_device **devices;
     };
 
     /**
