+++ src/thunderbots/software/radio_communication/mrf/libusb_errors.cpp
@@ -1,0 +1,39 @@
+#include "libusb_errors.h"
+
+#include <iostream>
+#include <string>
+#include <sstream>
+
+namespace
+{
+    std::string make_transfer_error_message(unsigned int endpoint, const char *msg)
+    {
+        std::stringstream ss;
+        ss << msg << " on " << ((endpoint & 0x80) ? "IN" : "OUT") << " endpoint "
+           << (endpoint & 0x7F) << std::endl;
+        return ss.str();
+    }
+
+}
+
+USB::Error::Error(const std::string &msg) : std::runtime_error(msg) {}
+
+USB::TransferError::TransferError(unsigned int endpoint, const char *msg)
+    : Error(make_transfer_error_message(endpoint, msg))
+{
+}
+
+USB::TransferTimeoutError::TransferTimeoutError(unsigned int endpoint)
+    : TransferError(endpoint, "Transfer timed out")
+{
+}
+
+USB::TransferStallError::TransferStallError(unsigned int endpoint)
+    : TransferError(endpoint, "Transfer stalled")
+{
+}
+
+USB::TransferCancelledError::TransferCancelledError(unsigned int endpoint)
+    : TransferError(endpoint, "Transfer cancelled")
+{
+}
